+++ src/device/tsp.script/Script/DisplaySubsystemExtensions/FirmwareManager-delete.cs
@@ -48,17 +48,24 @@ public static partial class FirmwareManager
         displaySubsystem.DisplayLine( 2, $"Deleting {script.Node.Number}:{script.Name}" );
 
         bool deleted = script.Node.IsController
-            ? FirmwareScript.ScriptNameExists( scriptNames, script.Name )
-                ? session.DeleteSavedScript( script.Name )
-                : session.NillScript( script.Name )
-            : FirmwareScript.ScriptNameExists( scriptNames, script.Name )
-                ? session.DeleteSavedScript( script.Node.Number, script.Name )
-                : session.NillScript( script.Node.Number, script.Name );
+             ? session.DeleteSavedScript( script.Name )
+             : session.DeleteSavedScript( script.Node.Number, script.Name );
+
+#if false
+        bool deleted = script.Node.IsController
+                ? FirmwareScript.ScriptNameExists( scriptNames, script.Name )
+                    ? session.DeleteSavedScript( script.Name )
+                    : session.NillScript( script.Name )
+                : FirmwareScript.ScriptNameExists( scriptNames, script.Name )
+                    ? session.DeleteSavedScript( script.Node.Number, script.Name )
+                    : session.NillScript( script.Node.Number, script.Name );
 
         // todo: check status byte for errors.
 
+#endif
         if ( !deleted )
             throw new InvalidOperationException( $"{session.ResourceNameNodeCaption} script '{script.Name}' still exists after nil." );
+
 
         // todo: check status byte for errors.