+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/NodeMethods_.cs
@@ -1,0 +1,8 @@
+namespace cc.isr.VI.Tsp.Script.SessionBaseExtensions;
+
+/// <summary>   A session base methods. </summary>
+/// <remarks>   2025-04-10. </remarks>
+[CLSCompliant( false )]
+public static partial class NodeMethods
+{
+}