+++ src/resource/pith/Syntax/Tsp/Lua.cs
@@ -45,14 +45,7 @@ public static class Lua
 
     #endregion
 
-    #region " script commands "
-
-    /// <summary>
-    /// Gets a command to retrieve a catalog from the local node.
-    /// This command must be enclosed in a 'do end' construct.
-    /// a print(names) or dataqueue.add(names) needs to be added to get the data through.
-    /// </summary>
-    public const string ScriptCatalogGetterCommand = "local names='' for name in script.user.catalog() do names = names .. name .. ',' end";
+    #region " string loader commands "
 
     /// <summary>   (Immutable) the load string LUA command. </summary>
     public const string LoadStringCommand = "loadstring";