+++ changelog.md
@@ -2,6 +2,16 @@
 All notable changes to these libraries will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [8.1.9233] - 2025-04-12
+- Resource Pith TSP Script Syntax: 
+  - Move find and catalog formats from Lua.
+- Resource Pith TSP Node Syntax: 
+  - Add catalog command format.
+  - add find saved script command.
+- Tsp Session Base Extensions
+  - Use TSP Syntax script constants.
+  - Add find and list saved script for nodes.
+
 ## [8.1.9232] - 2025-04-11
 - Pith Lua Syntax:
   - Rename ParseLuaChunkLine to ParseLuaChunkLine;
@@ -281,6 +291,6 @@ which has issued a defect report.
 
 &copy; 2012 Integrated Scientific Resources, Inc. All rights reserved.
 
-[8.1.9232]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9233]: https://www.github.com/atecoder/dn.vi.ivi
 [vs.VI]: https://www.github.com/atecoder/dn.vi.ivi
 [vs.Visa]: https://bitbucket.org/davidhary/vs.io.visa